Flyway done. Validation need to be check.

+++ src/main/java/travelagency/domain/Accommodation.java
@@ -24,5 +24,5 @@ public class Accommodation {
     @JoinColumn(name = "travel_id")
     private Travel travel;
 
-    private double price;
+    private Integer price;
 }
